Fix permission error

+++ src/main/java/me/melijn/jda/commands/management/PurgeCommand.java
@@ -30,14 +30,18 @@ public class PurgeCommand extends Command {
                 String[] args = event.getArgs().split("\\s+");
                 if (args.length == 1 && args[0].matches("^([1-9][0-9]{0,2}|1000)$")) {
                     int toPurgeAmount = Integer.parseInt(args[0]);
-                    event.getTextChannel().getIterableHistory().takeAsync(toPurgeAmount + 1).thenAccept(messages -> {
-                        messages.forEach(message -> MessageHelper.purgedMessageDeleter.putIfAbsent(message.getIdLong(), event.getAuthorId()));
-                        event.getTextChannel().purgeMessages(messages);
-                        event.getTextChannel().sendMessage("**Done**").queue(m -> {
-                            m.delete().queueAfter(3, TimeUnit.SECONDS);
-                            MessageHelper.botDeletedMessages.add(m.getIdLong());
+                    if (event.getMember().hasPermission(event.getTextChannel(), Permission.MESSAGE_HISTORY, Permission.MESSAGE_MANAGE)) {
+                        event.getTextChannel().getIterableHistory().takeAsync(toPurgeAmount + 1).thenAccept(messages -> {
+                            messages.forEach(message -> MessageHelper.purgedMessageDeleter.putIfAbsent(message.getIdLong(), event.getAuthorId()));
+                            event.getTextChannel().purgeMessages(messages);
+                            event.getTextChannel().sendMessage("**Done**").queue(m -> {
+                                m.delete().queueAfter(3, TimeUnit.SECONDS);
+                                MessageHelper.botDeletedMessages.add(m.getIdLong());
+                            });
                         });
-                    });
+                    } else {
+                        event.reply("I need the permission MESSAGE_HISTORY and MESSAGE_MANAGE in the TextChannel you want me to clear");
+                    }
                 } else {
                     MessageHelper.sendUsage(this, event);
                 }
